requirements.txt
requests
pandas
matplotlib
python-dotenv
web3
streamlit
plotly
